requirements.txt
torch
torchvision
torchcam
numpy==1.26.4
pillow==10.1.0
matplotlib==3.8.4
scikit-learn==1.4.0
opencv-python==4.9.0.80
tqdm==4.66.2
pennylane==0.35.0